# use a proper shared library for the core compiler intead of a ji file