This should be faster (ref https://github.com/JuliaLang/julia/commit/095fc99b69459d43c2ab3267ade9f491f52b92cc)

Just a random example (with the second step being the important one):

Before:
```
julia> PackageCompiler.create_sysimage(["Glob"]; sysimage_path="glob.dylib", incremental=false)
✔ [02m:27s] PackageCompiler: creating compiler .ji image (incremental=false)
✔ [01m:31s] PackageCompiler: compiling fresh sysimage (incremental=false)
✔ [00m:29s] PackageCompiler: compiling nonincremental system image
```

After:
```
julia> PackageCompiler.create_sysimage(["Glob"]; sysimage_path="glob.dylib", incremental=false)
✔ [02m:35s] PackageCompiler: creating compiler sysimage (incremental=false)
✔ [01m:05s] PackageCompiler: compiling fresh sysimage (incremental=false)
✔ [00m:28s] PackageCompiler: compiling nonincremental system image
```